python/requirements.txt
Pillow==5.3.0
pygame==1.9.4
unicorn-hat-sim==1.0.0
websockets==7.0
web.py=web.py-0.40.dev1